Moved profiling under benchmark and reduced rounds for benckmark tasks
