+++ Gemfile
@@ -72,7 +72,7 @@ end
 
 group :development do
   gem 'annotate', '~> 2.7'
-  gem 'ruby-prof'
+  gem 'ruby-prof', '~> 0.17.0', require: false
 end
 
 group :test do
